requirements.txt
pandas>=2.0.0
openpyxl>=3.1.2
psycopg2-binary>=2.9.9
python-magic>=0.4.27
